HUGE UPDATE
Added updating "Install Dependencies"
Added fb token check (to test if valid)
Included restart in Facebots.js
Removed useless duplicate dependencies
Separated CONFIG for "clean" access
Fixed RestartTimer to be off if set to 0 (when set to 0 it would restart
immediately)
Added Multiple fb accounts (NOT SIMULTANEOUS YET, USEFUL IF YOU HAVE
MULTIPLE COMPUTERS/VPS)
Updated README.md
Fix for real bot count
Now bots move and collect food and also split 2 times (MORE XP)
Console now gets cleared on restart
Added restarTimer notification every minute (console.log())
Fixed server.key which was malused and resulted in "undefined"
Fixed "ballAppear" which is the wrong event to listen to ( It gets
triggered by everything, even food) and set it to myNewBall
Removed old code
Many bugs fixed...
**NOTE: NOW MORE RAM/CPU WILL BE USED AS THE BOTS NOW MOVE AND COLLECT
FOOD (MORE XP)...**
...
