+++ Start (Windows).bat
@@ -1,0 +1,6 @@
+@echo off
+title ---- FaceBots ----
+cd src/
+call node Start
+echo Press any key to exit...
+pause >nul
