+++ src/Start.js
@@ -1,12 +1,11 @@
 var forever = require('forever-monitor');
-
 var child = new (forever.Monitor)('Facebots.js', {
     silent: false,
-    args: []
+    args: ["--spinSleepTime", "10000"]
 });
-
 child.on('exit', function () {
-    console.log('Facebots are restarting...');
+    process.stdout.write('\033c'); // Clear console
+    console.log('Facebots restarting...');
 });
 
 child.start();
