--- used in old version ( windows restart version )/Start (Windows).bat
@@ -1,8 +1,0 @@
-@echo off
-:start
-call node facebots
-cls
-echo Restarting...
-timeout /t 3 /nobreak > NUL
-goto start
-pause >nul
